LPS-165285 Inline CompositeFacetProcessor into DefaultFacetTranslator, since it's only used there
