+++ modules/apps/portal-search-elasticsearch7/portal-search-elasticsearch7-impl/src/main/java/com/liferay/portal/search/elasticsearch7/internal/facet/DefaultFacetTranslator.java
@@ -14,9 +14,11 @@
 
 package com.liferay.portal.search.elasticsearch7.internal.facet;
 
+import com.liferay.portal.kernel.json.JSONObject;
 import com.liferay.portal.kernel.search.BooleanClause;
 import com.liferay.portal.kernel.search.Query;
 import com.liferay.portal.kernel.search.facet.Facet;
+import com.liferay.portal.kernel.search.facet.config.FacetConfiguration;
 import com.liferay.portal.kernel.search.filter.BooleanFilter;
 import com.liferay.portal.kernel.search.filter.Filter;
 import com.liferay.portal.kernel.search.filter.FilterTranslator;
@@ -25,6 +27,7 @@ import com.liferay.portal.kernel.util.MapUtil;
 
 import java.util.ArrayList;
 import java.util.Collection;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
@@ -33,10 +36,17 @@ import org.elasticsearch.index.query.BoolQueryBuilder;
 import org.elasticsearch.index.query.QueryBuilder;
 import org.elasticsearch.index.query.QueryBuilders;
 import org.elasticsearch.search.aggregations.AggregationBuilder;
+import org.elasticsearch.search.aggregations.AggregationBuilders;
+import org.elasticsearch.search.aggregations.BucketOrder;
+import org.elasticsearch.search.aggregations.bucket.terms.IncludeExclude;
+import org.elasticsearch.search.aggregations.bucket.terms.TermsAggregationBuilder;
 import org.elasticsearch.search.builder.SearchSourceBuilder;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
+import org.osgi.service.component.annotations.ReferenceCardinality;
+import org.osgi.service.component.annotations.ReferencePolicy;
+import org.osgi.service.component.annotations.ReferencePolicyOption;
 
 /**
  * @author Michael C. Han
@@ -82,8 +92,7 @@ public class DefaultFacetTranslator implements FacetTranslator {
 				}
 			}
 
-			AggregationBuilder aggregationBuilder =
-				_facetProcessor.processFacet(facet);
+			AggregationBuilder aggregationBuilder = _processFacet(facet);
 
 			if (aggregationBuilder != null) {
 				AggregationBuilder postProcessAggregationBuilder =
@@ -115,6 +124,29 @@ public class DefaultFacetTranslator implements FacetTranslator {
 		return aggregationBuilder;
 	}
 
+	@Reference(
+		cardinality = ReferenceCardinality.MULTIPLE,
+		policy = ReferencePolicy.DYNAMIC,
+		policyOption = ReferencePolicyOption.GREEDY, target = "(class.name=*)"
+	)
+	protected void setFacetProcessor(
+		FacetProcessor<SearchRequestBuilder> facetProcessor,
+		Map<String, Object> properties) {
+
+		String className = MapUtil.getString(properties, "class.name");
+
+		_facetProcessors.put(className, facetProcessor);
+	}
+
+	protected void unsetFacetProcessor(
+		FacetProcessor<SearchRequestBuilder> facetProcessor,
+		Map<String, Object> properties) {
+
+		String className = MapUtil.getString(properties, "class.name");
+
+		_facetProcessors.remove(className);
+	}
+
 	private FacetProcessorContext _getFacetProcessorContext(
 		Collection<Facet> facets, boolean basicFacetSelection) {
 
@@ -139,6 +171,19 @@ public class DefaultFacetTranslator implements FacetTranslator {
 		return boolQueryBuilder;
 	}
 
+	private AggregationBuilder _processFacet(Facet facet) {
+		Class<?> clazz = facet.getClass();
+
+		FacetProcessor<SearchRequestBuilder> facetProcessor =
+			_facetProcessors.get(clazz.getName());
+
+		if (facetProcessor == null) {
+			facetProcessor = _defaultFacetProcessor;
+		}
+
+		return facetProcessor.processFacet(facet);
+	}
+
 	private QueryBuilder _translateBooleanClause(
 		BooleanClause<Filter> booleanClause) {
 
@@ -150,8 +195,50 @@ public class DefaultFacetTranslator implements FacetTranslator {
 		return _filterTranslator.translate(booleanFilter, null);
 	}
 
-	@Reference(service = CompositeFacetProcessor.class)
-	private FacetProcessor<SearchRequestBuilder> _facetProcessor;
+	private final FacetProcessor<SearchRequestBuilder> _defaultFacetProcessor =
+		new FacetProcessor<SearchRequestBuilder>() {
+
+			@Override
+			public AggregationBuilder processFacet(Facet facet) {
+				TermsAggregationBuilder termsAggregationBuilder =
+					AggregationBuilders.terms(
+						FacetUtil.getAggregationName(facet));
+
+				termsAggregationBuilder.field(facet.getFieldName());
+
+				FacetConfiguration facetConfiguration =
+					facet.getFacetConfiguration();
+
+				JSONObject dataJSONObject = facetConfiguration.getData();
+
+				String include = dataJSONObject.getString("include", null);
+
+				if (include != null) {
+					termsAggregationBuilder.includeExclude(
+						new IncludeExclude(include, null));
+				}
+
+				int minDocCount = dataJSONObject.getInt("frequencyThreshold");
+
+				if (minDocCount > 0) {
+					termsAggregationBuilder.minDocCount(minDocCount);
+				}
+
+				termsAggregationBuilder.order(BucketOrder.count(false));
+
+				int size = dataJSONObject.getInt("maxTerms");
+
+				if (size > 0) {
+					termsAggregationBuilder.size(size);
+				}
+
+				return termsAggregationBuilder;
+			}
+
+		};
+
+	private final Map<String, FacetProcessor<SearchRequestBuilder>>
+		_facetProcessors = new HashMap<>();
 
 	@Reference(target = "(search.engine.impl=Elasticsearch)")
 	private FilterTranslator<QueryBuilder> _filterTranslator;
